Added unitary tests and fixed bug in yaml process
Fix https://github.com/ubuntu/desktop-snaps/issues/19
